+++ src/controllers/index.ts
@@ -1,3 +1,3 @@
 export { default as errorController } from "./errorController";
-export { default as topPlayersController } from "./topPlayersController";
-export { default as topClubsController } from "./topClubsController";
+export { default as topPlayersController } from "./stat/playersController";
+export { default as topClubsController } from "./stat/clubsController";
