+++ src/models/index.ts
@@ -1,0 +1,4 @@
+export { default as PlayerData } from "./PlayerData";
+export { default as ClubData } from "./ClubData";
+export { default as topClubsCategory } from "./category/topClubsCategory";
+export { default as topPlayersCategory } from "./category/topPlayersCategory";
