+++ src/routes/index.ts
@@ -1,34 +1,34 @@
 import express, { Request, Response } from "express";
 import { topPlayersController, topClubsController } from "../controllers";
-import topClubsCategory from "../models/topClubsCategory";
-import topPlayersCategory from "../models/topPlayersCategory";
-
-const apiRouter: express.Router = express.Router();
-
-(function registerRoutes() {
-    apiRouter.get("/topPlayers/:category", topPlayersController);
-    apiRouter.get("/topClubs/:category", topClubsController);
-})();
+import topClubsCategory from "../models/category/topClubsCategory";
+import topPlayersCategory from "../models/category/topPlayersCategory";
 
 namespace API {
+    export const apiRouter: express.Router = express.Router();
+
+    (function registerRoutes() {
+        apiRouter.get("/stats/top/players/:category", topPlayersController);
+        apiRouter.get("/stats/top/clubs/:category", topClubsController);
+    })();
+
     export function controller(req: Request, res: Response) {
         return res.status(200).json({
             topPlayer: {
-                route: "/topPlayers/:category",
+                route: "/stats/top/players/:category",
                 params: {
                     category: topPlayersCategory,
                 },
             },
             topClub: {
-                route: "/topClubs/:category",
+                route: "/stats/top/players/:category",
                 params: {
                     categroy: topClubsCategory,
                 },
             },
         });
     }
+
+    apiRouter.get("", API.controller);
 }
 
-apiRouter.get("", API.controller);
-
-export default apiRouter;
+export default API.apiRouter;
