+++ src/scraper/stat/index.ts
@@ -1,0 +1,2 @@
+export { default as scrapePlayerData } from "./player";
+export { default as scrapeClubData } from "./club";
